+++ CHANGELOG.md
@@ -5,9 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Changed
-- Rename `enableAutoArchiveTranslations` into `enableArchivingTranslations`
-- Rename `disableAutoArchiveTranslations` into `disableArchivingTranslations`
+### Removed
+- The `auto archiving translations` feature.
 
 ## [0.6.0] - 2021-01-14
 ### Added